+++ io_scene_niftools/modules/nif_import/collision/havok.py
@@ -203,7 +203,7 @@ class BhkCollision(Collision):
             b_col_obj.nifcollision.max_linear_velocity = bhkshape.max_linear_velocity
             b_col_obj.nifcollision.max_angular_velocity = bhkshape.max_angular_velocity
 
-            # b_col_obj.nifcollision.oblivion_layer = NifFormat.OblivionLayer._enumkeys[bhkshape.layer]
+            b_col_obj.nifcollision.collision_layer = str(bhkshape.layer)
             # b_col_obj.nifcollision.quality_type = NifFormat.MotionQuality._enumkeys[bhkshape.quality_type]
             # b_col_obj.nifcollision.motion_system = NifFormat.MotionSystem._enumkeys[bhkshape.motion_system]
             # b_col_obj.nifcollision.col_filter = bhkshape.col_filter